+++ src/main/java/com/teamharmony/newscommunity/users/controller/UserController.java
@@ -6,10 +6,10 @@ import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.teamharmony.newscommunity.users.dto.UserResponseDto;
+import com.teamharmony.newscommunity.users.service.UserService;
 import com.teamharmony.newscommunity.users.vo.ProfileVO;
 import com.teamharmony.newscommunity.users.dto.SignupRequestDto;
 import com.teamharmony.newscommunity.users.entity.*;
-import com.teamharmony.newscommunity.users.service.UserService;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 import org.hibernate.exception.ConstraintViolationException;
